refactor: 패키지 구조 변경 및 컨벤션 정리
