+++ src/test/java/vendingmachine/domain/investmentmoney/InvestmentMoneyTest.java
@@ -1,4 +1,4 @@
-package vendingmachine.domain;
+package vendingmachine.domain.investmentmoney;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -6,6 +6,9 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+
+import vendingmachine.domain.investmentmoney.InvestmentMoney;
+import vendingmachine.domain.product.Product;
 
 class InvestmentMoneyTest {
