+++ src/test/java/vendingmachine/domain/possessioncoin/PossessionCoinsTest.java
@@ -1,4 +1,4 @@
-package vendingmachine.domain;
+package vendingmachine.domain.possessioncoin;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -7,6 +7,10 @@ import java.util.List;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import vendingmachine.domain.Coin;
+import vendingmachine.domain.change.Change;
+import vendingmachine.domain.investmentmoney.InvestmentMoney;
 
 class PossessionCoinsTest {
