+++ src/test/java/vendingmachine/domain/product/ProductTest.java
@@ -1,4 +1,4 @@
-package vendingmachine.domain;
+package vendingmachine.domain.product;
 
 import static org.assertj.core.api.Assertions.*;
